# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pangolin Catalog API
  description: Multi-tenant Apache Iceberg REST Catalog with advanced features including federated catalogs, credential vending, and fine-grained permissions
  contact:
    name: Pangolin API Support
    email: support@pangolin.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
paths:
...
  /api/v1/oauth/{provider}/authorize:
    get:
      tags:
      - OAuth
      summary: Initiate OAuth flow
      operationId: oauth_authorize
      parameters:
      - name: provider
        in: path
        description: OAuth provider (google, microsoft, github, okta)
        required: true
        schema:
          type: string
      responses:
        '302':
          description: Redirect to OAuth provider
        '400':
          description: Invalid provider
  /api/v1/oauth/{provider}/callback:
    get:
      tags:
      - OAuth
      summary: OAuth callback handler
      operationId: oauth_callback
      parameters:
      - name: provider
        in: path
        description: OAuth provider (google, microsoft, github, okta)
        required: true
        schema:
          type: string
      responses:
        '302':
          description: Redirect to frontend with token
        '400':
          description: Invalid provider or callback failed
        '500':
          description: Internal server error
...
  /api/v1/tenants:
    get:
      tags:
      - Tenants
      operationId: list_tenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantResponse'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Tenants
      operationId: create_tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '403':
          description: Forbidden in NO_AUTH mode
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
  /api/v1/tokens:
    post:
      tags:
      - Tokens
      summary: Generate a JWT token for a tenant
      description: This endpoint allows generating tokens for testing and development
      operationId: generate_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTokenRequest'
        required: true
      responses:
        '200':
          description: Token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTokenResponse'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /api/v1/users:
    get:
      tags:
      - Users
      summary: List all users (filtered by permissions)
      operationId: list_users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfo'
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
    post:
      tags:
      - Users
      summary: Create a new user
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    CreateTenantRequest:
      type: object
      required:
      - name
      properties:
        admin_password:
          type: string
          nullable: true
        admin_username:
          type: string
          nullable: true
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
          nullable: true
    CreateUserRequest:
      type: object
      description: Request to create a new user
      required:
      - username
      - email
      - role
      properties:
        email:
          type: string
        password:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        tenant_id:
          type: string
          format: uuid
          nullable: true
        username:
          type: string
...
    GenerateTokenRequest:
      type: object
      required:
      - tenant_id
      properties:
        expires_in_hours:
          type: integer
          format: int64
          nullable: true
          minimum: 0
        roles:
          type: array
          items:
            type: string
          nullable: true
        tenant_id:
          type: string
        username:
          type: string
          nullable: true
    GenerateTokenResponse:
      type: object
      required:
      - token
      - expires_at
      - tenant_id
      properties:
        expires_at:
          type: string
        tenant_id:
          type: string
        token:
          type: string
...
    OAuthProvider:
      type: string
      description: OAuth provider
      enum:
      - google
      - microsoft
      - github
      - okta
...
    TenantResponse:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
          nullable: true
...
    UserInfo:
      type: object
      description: Public user information
      required:
      - id
      - username
      - email
      - role
      properties:
        email:
          type: string
        id:
          type: string
          format: uuid
        oauth-provider:
          allOf:
          - $ref: '#/components/schemas/OAuthProvider'
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        tenant-id:
          type: string
          format: uuid
          nullable: true
        username:
          type: string
    UserRole:
      type: string
      description: User role in the system
      enum:
      - root
      - tenant-admin
      - tenant-user
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/users/login or /api/v1/tokens
...